ENH: Disable unnecessary compression support in LLVM build configuration
CMake building of llvm autodetects zstd and zlib from homebrew and automatically links against them.

A compiler warning is given:

```
ld: warning: building for macOS-11.0, but linking with dylib '/opt/homebrew/opt/zstd/lib/libzstd.1.dylib' which was built for newer version 15.0
```

Additionally, this prevents castxml from running on many systems due to the dependence on homebrew being installed with the same or newer version as was present during compilation.